Refactor and re-package custom exception classes
