+++ src/main/java/com/bytebreeze/quickdrop/exception/AlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.bytebreeze.quickdrop.exception.custom;
+package com.bytebreeze.quickdrop.exception;
 
 public class AlreadyExistsException extends RuntimeException {
