+++ src/main/java/com/bytebreeze/quickdrop/exception/FileValidationException.java
@@ -1,4 +1,4 @@
-package com.bytebreeze.quickdrop.exception.custom;
+package com.bytebreeze.quickdrop.exception;
 
 public class FileValidationException extends RuntimeException {
 	public FileValidationException(String message) {
