+++ src/main/java/com/bytebreeze/quickdrop/exception/ParcelNotFoundException.java
@@ -1,4 +1,4 @@
-package com.bytebreeze.quickdrop.exception.custom;
+package com.bytebreeze.quickdrop.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
