+++ src/main/java/com/bytebreeze/quickdrop/exception/SSLCommerzPaymentInitializationException.java
@@ -1,4 +1,4 @@
-package com.bytebreeze.quickdrop.exception.custom;
+package com.bytebreeze.quickdrop.exception;
 
 public class SSLCommerzPaymentInitializationException extends RuntimeException {
 	public SSLCommerzPaymentInitializationException(String message) {
