+++ src/main/java/com/bytebreeze/quickdrop/service/RiderService.java
@@ -10,10 +10,10 @@ import com.bytebreeze.quickdrop.entity.RiderEntity;
 import com.bytebreeze.quickdrop.enums.ParcelStatus;
 import com.bytebreeze.quickdrop.enums.Role;
 import com.bytebreeze.quickdrop.enums.VerificationStatus;
-import com.bytebreeze.quickdrop.exception.custom.AlreadyExistsException;
-import com.bytebreeze.quickdrop.exception.custom.ParcelAlreadyAssignedException;
-import com.bytebreeze.quickdrop.exception.custom.ParcelNotFoundException;
-import com.bytebreeze.quickdrop.exception.custom.UserNotFoundException;
+import com.bytebreeze.quickdrop.exception.AlreadyExistsException;
+import com.bytebreeze.quickdrop.exception.ParcelAlreadyAssignedException;
+import com.bytebreeze.quickdrop.exception.ParcelNotFoundException;
+import com.bytebreeze.quickdrop.exception.UserNotFoundException;
 import com.bytebreeze.quickdrop.mapper.RegisterRiderMapper;
 import com.bytebreeze.quickdrop.entity.ParcelEntity;
 import com.bytebreeze.quickdrop.repository.ParcelRepository;
