+++ src/main/java/com/bytebreeze/quickdrop/service/UserService.java
@@ -3,8 +3,8 @@ package com.bytebreeze.quickdrop.service;
 import com.bytebreeze.quickdrop.dto.request.UserProfileUpdateDto;
 import com.bytebreeze.quickdrop.dto.request.UserRegistrationRequestDTO;
 import com.bytebreeze.quickdrop.enums.Role;
-import com.bytebreeze.quickdrop.exception.custom.AlreadyExistsException;
-import com.bytebreeze.quickdrop.exception.custom.UserNotFoundException;
+import com.bytebreeze.quickdrop.exception.AlreadyExistsException;
+import com.bytebreeze.quickdrop.exception.UserNotFoundException;
 import com.bytebreeze.quickdrop.entity.UserEntity;
 import com.bytebreeze.quickdrop.repository.UserRepository;
 import com.bytebreeze.quickdrop.util.AuthUtil;
