+++ src/test/java/com/bytebreeze/quickdrop/controller/RiderControllerTest.java
@@ -10,8 +10,8 @@ import com.bytebreeze.quickdrop.dto.request.RiderRegistrationRequestDTO;
 import com.bytebreeze.quickdrop.dto.response.RiderViewCurrentParcelsResponseDTO;
 import com.bytebreeze.quickdrop.enums.ParcelStatus;
 import com.bytebreeze.quickdrop.enums.VerificationStatus;
-import com.bytebreeze.quickdrop.exception.custom.AlreadyExistsException;
-import com.bytebreeze.quickdrop.exception.custom.ParcelNotFoundException;
+import com.bytebreeze.quickdrop.exception.AlreadyExistsException;
+import com.bytebreeze.quickdrop.exception.ParcelNotFoundException;
 import com.bytebreeze.quickdrop.entity.ParcelEntity;
 import com.bytebreeze.quickdrop.entity.RiderEntity;
 import com.bytebreeze.quickdrop.service.ParcelService;
