+++ src/test/java/com/bytebreeze/quickdrop/service/FileStorageServiceTest.java
@@ -3,8 +3,8 @@ package com.bytebreeze.quickdrop.service;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-import com.bytebreeze.quickdrop.exception.custom.FileStorageException;
-import com.bytebreeze.quickdrop.exception.custom.FileValidationException;
+import com.bytebreeze.quickdrop.exception.FileStorageException;
+import com.bytebreeze.quickdrop.exception.FileValidationException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
