+++ src/test/java/com/bytebreeze/quickdrop/service/RiderServiceTest.java
@@ -7,7 +7,9 @@ import com.bytebreeze.quickdrop.dto.request.RiderOnboardingDTO;
 import com.bytebreeze.quickdrop.dto.request.RiderRegistrationRequestDTO;
 import com.bytebreeze.quickdrop.dto.response.*;
 import com.bytebreeze.quickdrop.enums.*;
-import com.bytebreeze.quickdrop.exception.custom.*;
+import com.bytebreeze.quickdrop.exception.AlreadyExistsException;
+import com.bytebreeze.quickdrop.exception.ParcelAlreadyAssignedException;
+import com.bytebreeze.quickdrop.exception.UserNotFoundException;
 import com.bytebreeze.quickdrop.mapper.RegisterRiderMapper;
 import com.bytebreeze.quickdrop.entity.*;
 import com.bytebreeze.quickdrop.repository.*;
